save_game-main/Backend/requirements.txt
anyio==3.6.1
asgiref==3.5.2
base58==2.1.1
certifi==2022.9.14
Django==4.1.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
ECPy==1.2.5
h11==0.12.0
httpcore==0.13.7
httpx==0.18.2
idna==3.4
psycopg2-binary==2.9.3
PyJWT==2.5.0
pytz==2022.2.1
rfc3986==1.5.0
sniffio==1.3.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
websockets==10.3
xrpl-py==1.6.0
